Fixed a refactored call where I was fetching the JNI-specific "thread identifier" field (getTID()) instead of fetching the "identity" field from the ICMP header. This was causing all ICMP operations to fail when invoked from within the RequestTracker.
Also fixed some problems with using TimeUnit to convert values for the
RTT of the pings.

+++ opennms-icmp/opennms-icmp-jicmp/src/main/java/org/opennms/netmgt/ping/PingRequest.java
@@ -56,11 +56,11 @@ final public class PingRequest extends AbstractPingRequest<IcmpSocket> {
         super(addr, tid, sequenceId, timeout, retries, logger, cb);
     }
     
-    PingRequest(InetAddress addr, long tid, short sequenceId, long timeout, int retries, PingResponseCallback cb) {
+    PingRequest(InetAddress addr, long tid, int sequenceId, long timeout, int retries, PingResponseCallback cb) {
         this(addr, tid, sequenceId, timeout, retries, ThreadCategory.getInstance(PingRequest.class), cb);
     }
     
-    PingRequest(InetAddress addr, short sequenceId, long timeout, int retries, PingResponseCallback cb) {
+    PingRequest(InetAddress addr, int sequenceId, long timeout, int retries, PingResponseCallback cb) {
         this(addr, s_nextTid++, sequenceId, timeout, retries, cb);
     }
